Added Drop menus
I was having an issue with this until i realized i placed them in the wrong container. The only thing i dont like is that the drop down menus carry over the "border-right: 1px solid white" i want to get rid of it but cant seem to figure it out at the moment. Its not a huge deal but i dont like it lol